Newton Analysis
look at Many Magnetic Displacement Newtons. You can see that the zeros of the displacement, which are the fixed points of newton's method, correctly have only one fixed point at the correct value on the interval. however, the increasingly large peaks (with high 2nd derivatives) indicate increasingly worse behaviour in the middle zones between eigenvalues. This is a really, really bad problem - it's more obvious how bad this is when one realizes the y-axis is actually ~8 times greater scale than the x.