Wikipedia maybe wrong? Ad hoc deviation works for Bessel.
Wikipedia's sum in the recurrence relation of the coefficients for the frobenius method gets wrong values. taking only the first (or, depending on ordering of the sum, last. this is purely ordering choice and not an actual term change) value of the sum gets the right answer

For the Bessel Function of order 1/2, correct first 3 coefficients are obtained with very little error [remember that the first was manually provided]. after that, values become very wrong, but this is reasonable and not a cause for worry probably.

The weird part is that the legendre polynomial coefficients are just wrong, unlike the bessel ones. Further testing on other known solution DEs is required.